BUG FIX: Fixed a bunch of stuff to get dockerized unit testing up and running with codeception
