+++ codeception.dist.yml
@@ -17,4 +17,7 @@ extensions:
         - Codeception\Command\GenerateWPCanonical
         - Codeception\Command\GenerateWPXMLRPC
 params:
-    - .env.testing
+    - tests/_envs/.env.utilities
+settings:
+  lint: true
+  colors: true
