+++ tests/_envs/.env.utilities
@@ -6,15 +6,15 @@ TEST_SITE_DB_HOST="database"
 TEST_SITE_DB_PORT="3306"
 TEST_SITE_DB_USER="wordpress"
 TEST_SITE_DB_PASSWORD="wordpress"
-TEST_SITE_TABLE_PREFIX="wp_"
-TEST_SITE_WP_URL="http://localhost"
+TEST_SITE_TABLE_PREFIX="test_"
+TEST_SITE_WP_URL="http://localhost:7253"
 TEST_SITE_WP_DOMAIN="localhost"
-TEST_SITE_ADMIN_EMAIL="admin@localhost"
+TEST_SITE_ADMIN_EMAIL="thomas@eighty20results.com"
 TEST_SITE_ADMIN_USERNAME="admin"
 TEST_SITE_ADMIN_PASSWORD="admin"
 
 TEST_DB_NAME="wordpress"
 TEST_DB_HOST="database"
 TEST_DB_USER="wordpress"
-TEST_DB_PASSWORD="password"
-TEST_TABLE_PREFIX="wp_"
+TEST_DB_PASSWORD="wordpress"
+TEST_TABLE_PREFIX="test_"
