+++ tests/wpunit/bootstrap.php
@@ -1,25 +1,5 @@
 <?php
 /*
- * *
- *   * Copyright (c) 2021. - Eighty / 20 Results by Wicked Strong Chicks.
- *   * ALL RIGHTS RESERVED
- *   *
- *   * This program is free software: you can redistribute it and/or modify
- *   * it under the terms of the GNU General Public License as published by
- *   * the Free Software Foundation, either version 3 of the License, or
- *   * (at your option) any later version.
- *   *
- *   * This program is distributed in the hope that it will be useful,
- *   * but WITHOUT ANY WARRANTY; without even the implied warranty of
- *   * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *   * GNU General Public License for more details.
- *   *
- *   * You should have received a copy of the GNU General Public License
- *   * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-/*
  *  Copyright (c) 2021. - Eighty / 20 Results by Wicked Strong Chicks.
  *  ALL RIGHTS RESERVED
  *
@@ -53,11 +33,11 @@ if ( ! defined( 'ABSPATH' ) ) {
 }
 // define fake PLUGIN_ABSPATH
 if ( ! defined( 'PLUGIN_ABSPATH' ) ) {
-	define( 'PLUGIN_ABSPATH', sys_get_temp_dir() . '/wp-content/plugins/e20r-members-list/' );
+	define( 'PLUGIN_ABSPATH', sys_get_temp_dir() . '/wp-content/plugins/00-e20r-utilities/' );
 }
 
 if ( ! defined( 'PLUGIN_PATH' ) ) {
-	define( 'PLUGIN_PATH', __DIR__ . '/../Utilities/' );
+	define( 'PLUGIN_PATH', __DIR__ . '/../00-e20r-utilities/' );
 }
 
 require_once __DIR__ . '/../../inc/autoload.php';
